Adding and modifying files from Angular tutorial
See the following website:
`https://jasonwatmore.com/post/2020/04/28/angular-9-user-registration-
and-login-example-tutorial#account-login-component-html`

This commit represents an initial attempt at
adding User Registration and Login files to this app.
It is NOT connected to the Rails backend yet,
and so far is mostly for learning Angular's structure and note taking.
More to come in later commits.